same for sponsors page
