+++ src/pages/Events/Events.tsx
@@ -4,31 +4,6 @@ import Divider from "../MainPage/Divider";
 import upcomingEvents from "./UpcomingEvents";
 import pastEvents from "./PastEvents";
 const Events = () => {
-  const pastEventLogos = [
-    "/event-logos/eotg-event-wi25.jpg",
-    "/event-logos/decaf-event-fa24.jpg",
-    "/event-logos/eotg-event-fa24.jpg",
-    "/event-logos/decaf-event-fa23.svg",
-    "/event-logos/sd-hacks-event-wi21.png",
-    "/event-logos/decaf-event-wi21.svg",
-    "/event-logos/eotg-event-wi21.svg",
-    "/event-logos/baking-event-fa20.png",
-    "/event-logos/matcha-event-fa20.svg",
-    "/event-logos/destress-event-fa20.svg",
-    "/event-logos/eotg-event-fa20.svg"
-  ];
-  const pastEventLinks = [
-    "https://tescucsd.org/decaf-fa24",
-    "https://tescucsd.org/eotg-fa24",
-    "https://decafucsd.com",
-    "https://www.eotgucsd.com",
-    "https://decafucsd.com",
-    "https://www.eotgucsd.com",
-    "https://decafucsd.com",
-    "https://www.eotgucsd.com",
-    "https://decafucsd.com",
-    "https://www.eotgucsd.com"
-  ];
   return (
     <>
       <img
