+++ src/pages/MainPage/Sponsors-info.tsx
@@ -1,0 +1,40 @@
+import L3Harris_Logo from "/logos/L3Harris_Logo.png";
+import Dynovas from "/logos/Dynovas Matrix Logo (Final).jpg";
+import LLNLlogoBLACK from "/logos/LLNLlogoBLACK.jpg";
+import Viasat_2018_Gradient_Web from "/logos/Viasat_2018_Gradient_Web (1).jpg";
+import LANL_Logo_Temp from "/logos/LANL_Logo_Temp.png";
+
+const sponsors = [
+  {
+    id: 1,
+    logo: L3Harris_Logo,
+    alt: "L3Harris",
+    website: "https://www.l3harris.com/"
+  },
+  {
+    id: 2,
+    logo: Dynovas,
+    alt: "Dynovas",
+    website: "https://www.dynovas.com/"
+  },
+  {
+    id: 3,
+    logo: LLNLlogoBLACK,
+    alt: "LLNL",
+    website: "https://www.llnl.gov/"
+  },
+  {
+    id: 4,
+    logo: Viasat_2018_Gradient_Web,
+    alt: "Viasat",
+    website: "https://www.viasat.com/"
+  },
+  {
+    id: 5,
+    logo: LANL_Logo_Temp,
+    alt: "LANL",
+    website: "https://www.lanl.gov/"
+  }
+];
+
+export default sponsors;
